Added 'plots.py' and the folder 'images' for emission models visualization
Added a new script 'plots.py' that will contain functions to generate plots.
Currently, it contains a function for generating a plot of emitted radiation models.
The script calculates and plots the linear and black body radiation models using the 'stochastic_resonance' module.
The resulting plot is saved in the 'images' directory as 'emitted_radiation_plot.png'. The 'images' directory will store the generated images.