Added function to plot F(T) using emission models in 'plots.py' file
Implemented the 'Plot_F()' function in 'plots.py' to plot the behavior of F(T) using the two different models for emitted radiation.
This function calculates and visualizes F(T) using the linear and black body models provided by the 'stochastic_resonance' module.
The resulting plot is saved in the 'images' directory as 'F_plot.png'.