Added Simulation and Data Storage of Temperature and Time Dynamics to 'simulation.py' Module
In this commit, I introduced simulations of temperature dynamics over time in the 'simulation.py' script.
The simulations are performed using the 'simulate_ito' function from the 'stochastic_resonance.py' module.
These simulations were conducted for two distinct levels of noise intensity, corresponding to different variance values.
For each variance value, a series of 10 independent simulations were executed.

The outcomes of these simulations are saved within the 'Temperature' array and the 'Time' array.
The 'Temperature' array has dimensions of len(V), num_simulations, num_steps; while the 'Time' array has dimensions of len(V), num_steps.
Here, the array 'V' encompasses the variance values utilized in the simulations — specifically, two values in this context.

Subsequently, these arrays are stored within the newly created 'data' folder.

The 'data' folder is not uploaded to the repository due to the large size of the files it contains.
However, the 'data' folder is present in the local directory.
This decision was made considering the substantial file sizes of the simulation outputs.