Introduced SNR_plot() function in 'plots.py' module
This commit implements the SNR_plot() function within the 'plots.py' module.
The function is designed to generate a plot depicting the relationship between Signal-to-Noise Ratio (SNR) and noise variance.
It relies on the variables 'V' and 'SNR' from the 'simulation.py' module.
The resulting plot is then stored as 'SNR.png' in the 'images' folder.