Added tests for simulate_ito_combinations_and_collect_results function
This commit adds a set of test functions to ensure the correct behavior of the `simulate_ito_combinations_and_collect_results` function.
The tests cover various aspects, including the length and type of returned time and temperature values, as well as the length and shape of time and temperature sequences.

Tests Included:
- Test if the time returned has the correct length (4).
- Test if the temperature returned has the correct length (4).
- Test if the time returned is of type list.
- Test if the temperature returned is of type list.
- Test if the time sequences returned have the correct length (num_steps = 100).
- Test if the temperature sequences returned have the correct shape (num_simulations = 1, num_steps = 100).